Add MuiSvgCuxtomIcon, add tooggle light/dark theme, add demo for Badges

+++ core/src/main/scala/io/kinoplan/scalajs/react/material/ui/core/styles/BreakpointValues.scala
@@ -5,20 +5,20 @@ import scala.scalajs.js.JSConverters._
 
 @js.native
 trait BreakpointValues extends js.Object {
-  def xs: Double = js.native
-  def sm: Double = js.native
-  def md: Double = js.native
-  def lg: Double = js.native
-  def xl: Double = js.native
+  def xs: Int = js.native
+  def sm: Int = js.native
+  def md: Int = js.native
+  def lg: Int = js.native
+  def xl: Int = js.native
 }
 
 object BreakpointValues {
   def apply(
-    xs: Double,
-    sm: Double,
-    md: Double,
-    lg: Double,
-    xl: Double
+    xs: Int,
+    sm: Int,
+    md: Int,
+    lg: Int,
+    xl: Int
   ) = {
     val o: Map[String, Any] = Map(
       "xs" -> xs,
